requirements.txt
scikit-learn==1.6.0
joblib
fastapi
uvicorn
numpy
pandas
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
datetime